+++ built/jtree.browser.js
@@ -2659,7 +2659,7 @@ class GrammarBackedRootNode extends GrammarBackedNode {
         // returns a report on what nodeTypes from its language the program uses
         const usage = new TreeNode();
         const grammarProgram = this.getGrammarProgramRoot();
-        grammarProgram.getConcreteAndAbstractNodeTypeDefinitions().forEach(def => {
+        grammarProgram.getValidConcreteAndAbstractNodeTypeDefinitions().forEach(def => {
             usage.appendLine([def.getNodeTypeIdFromDefinition(), "line-id", GrammarConstants.nodeType, def.getRequiredCellTypeIds().join(" ")].join(" "));
         });
         this.getTopDownArray().forEach((node, lineNumber) => {
@@ -2672,6 +2672,9 @@ class GrammarBackedRootNode extends GrammarBackedNode {
         return this.getTopDownArray()
             .map(child => child.getIndentation() + child.getLineHighlightScopes())
             .join("\n");
+    }
+    getCatchAllNodeConstructor(line) {
+        return GrammarBackedBlobNode;
     }
     getInPlaceCellTypeTreeWithNodeConstructorNames() {
         return this.getTopDownArray()
@@ -3515,9 +3518,12 @@ class AbstractGrammarDefinitionNode extends AbstractExtendibleTreeNode {
     getNodeTypeIdFromDefinition() {
         return this.getWord(1);
     }
-    // todo: remove. just reused nodeTypeId
+    // todo: remove? just reused nodeTypeId
     _getGeneratedClassName() {
         return this.getNodeTypeIdFromDefinition();
+    }
+    _hasValidNodeTypeId() {
+        return !!this._getGeneratedClassName();
     }
     _isAbstract() {
         return this.has(GrammarConstants.abstract);
@@ -3680,8 +3686,8 @@ class AbstractGrammarDefinitionNode extends AbstractExtendibleTreeNode {
                 .replace(/\\/g, "\\\\"))}\`)
         return this._cachedGrammarProgramRoot
       }`);
-            const defs = this.getLanguageDefinitionProgram().getConcreteAndAbstractNodeTypeDefinitions();
-            const nodeTypeMap = defs
+            const nodeTypeMap = this.getLanguageDefinitionProgram()
+                .getValidConcreteAndAbstractNodeTypeDefinitions()
                 .map(def => {
                 const id = def.getNodeTypeIdFromDefinition();
                 return `"${id}": ${id}`;
@@ -3838,7 +3844,7 @@ class GrammarProgram extends AbstractGrammarDefinitionNode {
     getErrorsInGrammarExamples() {
         const programConstructor = this.getRootConstructor();
         const errors = [];
-        this.getConcreteAndAbstractNodeTypeDefinitions().forEach(def => def.getExamples().forEach(example => {
+        this.getValidConcreteAndAbstractNodeTypeDefinitions().forEach(def => def.getExamples().forEach(example => {
             const exampleProgram = new programConstructor(example.childrenToString());
             exampleProgram.getAllErrors(example._getLineNumber() + 1).forEach(err => {
                 errors.push(err);
@@ -3863,7 +3869,7 @@ class GrammarProgram extends AbstractGrammarDefinitionNode {
     }
     getNodeTypeFamilyTree() {
         const tree = new TreeNode();
-        Object.values(this.getConcreteAndAbstractNodeTypeDefinitions()).forEach(node => {
+        Object.values(this.getValidConcreteAndAbstractNodeTypeDefinitions()).forEach(node => {
             const path = node.getAncestorNodeTypeIdsArray().join(" ");
             tree.touchNode(path);
         });
@@ -3878,16 +3884,23 @@ class GrammarProgram extends AbstractGrammarDefinitionNode {
     getLanguageDefinitionProgram() {
         return this;
     }
-    getConcreteAndAbstractNodeTypeDefinitions() {
-        return this.getChildrenByNodeConstructor(NonRootNodeTypeDefinition);
+    getValidConcreteAndAbstractNodeTypeDefinitions() {
+        return (this.getChildrenByNodeConstructor(NonRootNodeTypeDefinition).filter((node) => node._hasValidNodeTypeId()));
     }
     _getRootNodeTypeDefinitionNode() {
         if (!this._cache_rootNodeTypeNode) {
             this.forEach(def => {
-                if (def.has(GrammarConstants.root))
+                if (def.has(GrammarConstants.root) && def._hasValidNodeTypeId())
                     this._cache_rootNodeTypeNode = def;
             });
         }
+        // By default, have a very permissive basic root node.
+        if (!this._cache_rootNodeTypeNode)
+            this._cache_rootNodeTypeNode = this.concat(`${GrammarConstants.nodeType} defaultNodeType
+ ${GrammarConstants.root}
+ ${GrammarConstants.catchAllNodeType} GrammarBackedBlobNode
+${GrammarConstants.nodeType} GrammarBackedBlobNode
+ ${GrammarConstants.baseNodeType} ${GrammarConstants.blobNode}`)[0];
         return this._cache_rootNodeTypeNode;
     }
     getExtensionName() {
@@ -3943,11 +3956,10 @@ class GrammarProgram extends AbstractGrammarDefinitionNode {
         return TreeUtils.ucfirst(this.getExtensionName());
     }
     _rootNodeDefToJavascriptClass(jtreePath, forNodeJs = true) {
-        const defs = this.getConcreteAndAbstractNodeTypeDefinitions();
+        const defs = this.getValidConcreteAndAbstractNodeTypeDefinitions();
         // todo: throw if there is no root node defined
-        const rootNode = this._getRootNodeTypeDefinitionNode();
         const nodeTypeClasses = defs.map(def => def._nodeDefToJavascriptClass()).join("\n\n");
-        const rootName = rootNode._getGeneratedClassName();
+        const rootName = this._getRootNodeTypeDefinitionNode()._getGeneratedClassName();
         let exportScript = "";
         if (forNodeJs) {
             exportScript = `module.exports = ${rootName};
@@ -3973,7 +3985,7 @@ ${exportScript}
         const variables = Object.keys(cellTypeDefs)
             .map(name => ` ${name}: '${cellTypeDefs[name].getRegexString()}'`)
             .join("\n");
-        const defs = this.getConcreteAndAbstractNodeTypeDefinitions().filter(kw => !kw._isAbstract());
+        const defs = this.getValidConcreteAndAbstractNodeTypeDefinitions().filter(kw => !kw._isAbstract());
         const nodeTypeContexts = defs.map(def => def.getMatchBlock()).join("\n\n");
         const includes = defs.map(nodeTypeDef => `  - include: '${nodeTypeDef.getNodeTypeIdFromDefinition()}'`).join("\n");
         return `%YAML 1.2
@@ -3990,17 +4002,6 @@ contexts:
 ${includes}
 
 ${nodeTypeContexts}`;
-    }
-    // A language where anything goes.
-    // todo: can we remove? can we make the default language not require any grammar node?
-    static getTheAnyLanguageRootConstructor() {
-        return new GrammarProgram(`${GrammarConstants.nodeType} anyLanguage
- ${GrammarConstants.root}
- ${GrammarConstants.catchAllNodeType} anyNode
-${GrammarConstants.nodeType} anyNode
- ${GrammarConstants.catchAllCellType} anyWord
- ${GrammarConstants.firstCellType} anyWord
-${GrammarConstants.cellType} anyWord`).getRootConstructor();
     }
 }
 GrammarProgram._languages = {};
@@ -4465,7 +4466,7 @@ jtree.TreeNode = TreeNode;
 jtree.GrammarProgram = GrammarProgram;
 jtree.UnknownGrammarProgram = UnknownGrammarProgram;
 jtree.TreeNotationCodeMirrorMode = TreeNotationCodeMirrorMode;
-jtree.getVersion = () => "33.0.2";
+jtree.getVersion = () => "34.0.0";
 window.jtree
     = jtree;
 class Upgrader extends TreeNode {